UI API cleanup and UI return format tests.
